+++ api/build.gradle.kts
@@ -1,10 +1,4 @@
-object Versions {
-    const val COROUTINE_VERSION = "1.8.1"
-    const val JACKSON_KOTLIN = "2.17.0"
-    const val KOTLIN_JSDL = "3.5.3"
-    const val TEST_CONTAINER = "1.19.1"
-    const val LOGGER = "7.0.0"
-}
+import org.contribhub.Versions
 
 plugins {
     kotlin("plugin.spring") version "1.9.25"
