+++ buildSrc/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    id("org.gradle.kotlin.kotlin-dsl") version "4.2.1"
+}
+
+repositories {
+    mavenCentral()
+}
