+++ buildSrc/src/main/kotlin/org/contribhub/Versions.kt
@@ -1,0 +1,9 @@
+package org.contribhub
+
+object Versions {
+    const val COROUTINE_VERSION = "1.8.1"
+    const val JACKSON_KOTLIN = "2.17.0"
+    const val KOTLIN_JSDL = "3.5.3"
+    const val TEST_CONTAINER = "1.19.1"
+    const val LOGGER = "7.0.0"
+}
